# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Pet Manager API
  version: 0.1.0
  description: |
    REST API of Pet Manager. All endpoints are under `/api`.

    **Authentication:** send `Authorization: Bearer <token>` (from `/auth/login`
    or `/auth/register`). Data endpoints also work without a token when the
    instance runs in anonymous mode (see README). `/admin/*` always requires an
    admin token, `/household/*` a real (non-anonymous) login.

    **Households:** every account and every pet belongs to one household. All
    pet data is scoped to the caller's household; pets of other households
    answer with `404 PET_NOT_FOUND`.

    **Errors** always have the shape `{ "error": { "code", "message", "details" } }`.
    `code` is stable (e.g. `AUTH_INVALID_CREDENTIALS`, `VALIDATION_FAILED`,
    `PET_NOT_FOUND`) and is translated by the client.

    Dates are `YYYY-MM-DD`, timestamps ISO 8601 with offset. Empty strings in
    optional text fields are stored as `null`.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api
security:
  - bearer: []
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register (the first account becomes admin)
      description: >-
        Rate limited. The first account becomes admin and gets the first household.
        With `invite_code` the account joins that household (works even when
        registration is closed). Otherwise a new household is created, which is
        only possible if an admin allowed registration.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, display_name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 10, description: max. 72 bytes }
                display_name: { type: string, maxLength: 100 }
                locale: { $ref: '#/components/schemas/Locale' }
                household_name:
                  { type: string, maxLength: 100, description: Name of the new household }
                invite_code: { type: string, description: Join an existing household }
      responses:
        '201': { $ref: '#/components/responses/Session' }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /auth/login:
    post:
      tags: [auth]
      summary: Log in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Session' }
        '401': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /auth/me:
    get:
      tags: [auth]
      summary: 'Current user (or `{ anonymous: true }` in anonymous mode)'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/User'
                  - type: object
                    properties:
                      { anonymous: { type: boolean, const: true }, role: { type: string } }
    patch:
      tags: [auth]
      summary: Update own profile
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name: { type: string }
                locale: { $ref: '#/components/schemas/Locale' }
      responses:
        '200':
          {
            description: OK,
            content: { application/json: { schema: { $ref: '#/components/schemas/User' } } },
          }
  /auth/password:
    post:
      tags: [auth]
      summary: Change own password (signs out all other sessions)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string }
                new_password: { type: string, minLength: 10 }
      responses:
        '200':
          description: New token for the current session
          content:
            {
              application/json:
                { schema: { type: object, properties: { token: { type: string } } } },
            }
        '400': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Session:
      description: Logged in
      content:
        application/json:
          schema:
            type: object
            properties:
              token: { type: string }
              user: { $ref: '#/components/schemas/User' }
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Locale: { type: string, enum: [en, de] }
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [VALIDATION_FAILED] }
            message: { type: string }
            details:
              type: array
              items:
                {
                  type: object,
                  properties:
                    { path: { type: string }, code: { type: string }, message: { type: string } },
                }
    User:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        display_name: { type: string }
        role: { type: string, enum: [admin, user] }
        locale: { $ref: '#/components/schemas/Locale' }
        household_id: { type: integer }
...